Reflect actual arch of found man page in rel_canonical_url
Extract the actual arch of the page from its path as already
stored in LASTMODFILE.

We don't redirect the browser to the actual arch as we want to
keep the user-selected arch intact as much as possible. It is
already lost upon traversing links like "x86/boot(8)".

Remove any port number in rel_canonical_url from the URL on
the basis that production deployment is on a standard port
and only that instance is generally crawled by search engines.